refactor GenericJourney
