+++ src/test/scala/org/kibanaLoadTest/simulation/generic/GenericJourney.scala
@@ -16,6 +16,7 @@ import org.kibanaLoadTest.helpers.{ESArchiver, Helper, HttpHelper, KbnClient}
 import org.kibanaLoadTest.simulation.generic
 import org.kibanaLoadTest.simulation.generic.mapping.{Journey, Step, TestData}
 import org.kibanaLoadTest.simulation.generic.mapping.JourneyJsonProtocol._
+import org.kibanaLoadTest.simulation.generic.core.{ApiCall, JourneyBuilder}
 import org.slf4j.{Logger, LoggerFactory}
 import spray.json._
 
@@ -27,218 +28,8 @@ import scala.collection.mutable.ListBuffer
 import scala.io.Source._
 import scala.util.Using
 
-object ApiCall {
-  def execute(
-      requests: List[mapping.Request],
-      config: KibanaConfiguration
-  ): ChainBuilder = {
-    // Workaround for https://github.com/gatling/gatling/issues/3783
-    val parent :: children = requests
-    val httpParentRequest = httpRequest(parent.http, config)
-    if (children.isEmpty) {
-      exec(httpParentRequest)
-    } else {
-      val childHttpRequests: Seq[HttpRequestBuilder] =
-        (children.map(request => httpRequest(request.http, config))).toSeq
-      exec(httpParentRequest.resources(childHttpRequests: _*))
-    }
-  }
-
-  def httpRequest(
-      request: mapping.Http,
-      config: KibanaConfiguration
-  ): HttpRequestBuilder = {
-    val excludeHeaders = List(
-      "Content-Length",
-      "Kbn-Version",
-      "Traceparent",
-      "Authorization",
-      "Cookie",
-      "X-Kbn-Context"
-    )
-    val defaultHeaders = request.headers.--(excludeHeaders.iterator)
-
-    var headers = defaultHeaders
-    if (request.headers.contains("Kbn-Version")) {
-      headers += ("Kbn-Version" -> config.buildVersion)
-    }
-    if (request.headers.contains("Cookie")) {
-      headers += ("Cookie" -> "#{Cookie}")
-    }
-    val requestName = s"${request.method} ${request.path
-      .replaceAll(".+?(?=\\/bundles)", "") + request.query.getOrElse("")}"
-    val url = request.path + request.query.getOrElse("")
-    request.method match {
-      case "GET" =>
-        http(requestName)
-          .get(url)
-          .headers(headers)
-          .check(status.is(request.statusCode))
-      case "POST" =>
-        request.body match {
-          case Some(value) =>
-            // https://gatling.io/docs/gatling/reference/current/http/request/#stringbody
-            // Gatling uses #{value} syntax to pass session values, we disable it by replacing # with ##
-            // $ was deprecated, but Gatling still identifies it as session attribute
-            val bodyString = value.replaceAll("\\$\\{\\{", "{{")
-            http(requestName)
-              .post(url)
-              .body(StringBody(bodyString))
-              .asJson
-              .headers(headers)
-              .check(status.is(request.statusCode))
-          case _ =>
-            http(requestName)
-              .post(url)
-              .asJson
-              .headers(headers)
-              .check(status.is(request.statusCode))
-        }
-
-      case "PUT" =>
-        http(requestName)
-          .put(url)
-          .headers(headers)
-          .check(status.is(request.statusCode))
-      case "DELETE" =>
-        http(requestName)
-          .delete(url)
-          .headers(headers)
-          .check(status.is(request.statusCode))
-      case _ =>
-        throw new IllegalArgumentException(s"Invalid method ${request.method}")
-    }
-  }
-}
-
 class GenericJourney extends Simulation {
   val logger: Logger = LoggerFactory.getLogger("GenericJourney")
-  val modelsMap = Map(
-    "constantConcurrentUsers" -> "closed",
-    "rampConcurrentUsers" -> "closed",
-    "atOnceUsers" -> "open",
-    "rampUsers" -> "open",
-    "rampUsersPerSec" -> "open",
-    "constantUsersPerSec" -> "open",
-    "stressPeakUsers" -> "open"
-  )
-
-  def getDuration(duration: String) = {
-    duration.takeRight(1) match {
-      case "s" => duration.dropRight(1).toInt
-      case "m" => duration.dropRight(1).toInt * 60
-      case _ =>
-        throw new IllegalArgumentException(
-          s"Invalid duration format: ${duration}"
-        )
-    }
-  }
-
-  def scenarioSteps(
-      journey: Journey,
-      config: KibanaConfiguration
-  ): ChainBuilder = {
-    var steps: ChainBuilder = exec()
-    var priorStream: Option[mapping.RequestStream] =
-      Option.empty
-    for (stream <- journey.streams) {
-      val priorDate =
-        priorStream.map(_.endTime).getOrElse(stream.endTime)
-      val pauseDuration =
-        Math.max(0L, stream.startTime.getTime - priorDate.getTime)
-      if (pauseDuration > 0L) {
-        steps = steps.pause(pauseDuration.toString, TimeUnit.MILLISECONDS)
-      }
-
-      if (!stream.requests.isEmpty) {
-        steps = steps.exec(ApiCall.execute(stream.requests, config))
-        priorStream = Option(stream)
-      }
-    }
-    steps
-  }
-
-  def scenarioForStage(
-      steps: ChainBuilder,
-      scenarioName: String
-  ): ScenarioBuilder = {
-    scenario(scenarioName).exec(steps)
-  }
-
-  def closedModelStep(step: Step): ClosedInjectionStep = {
-    logger.info(s"Closed model: building ${step.toString}")
-    step.action match {
-      case "constantConcurrentUsers" =>
-        constantConcurrentUsers(step.userCount.get) during (getDuration(
-          step.duration.get
-        ))
-      case "rampConcurrentUsers" =>
-        rampConcurrentUsers(
-          step.minUsersCount.get
-        ) to step.maxUsersCount.get during (getDuration(step.duration.get))
-      case _ =>
-        throw new IllegalArgumentException(
-          s"Invalid closed model: ${step.action}"
-        )
-    }
-  }
-
-  def openModelStep(step: Step): OpenInjectionStep = {
-    logger.info(s"Open model: building${step.toString}")
-    step.action match {
-      case "atOnceUsers" =>
-        atOnceUsers(step.userCount.get)
-      case "rampUsers" =>
-        rampUsers(step.userCount.get).during(getDuration(step.duration.get))
-      case "constantUsersPerSec" =>
-        constantUsersPerSec(step.userCount.get.toDouble)
-          .during(getDuration(step.duration.get))
-      case "stressPeakUsers" =>
-        stressPeakUsers(step.userCount.get)
-          .during(getDuration(step.duration.get))
-      case "rampUsersPerSec" =>
-        rampUsersPerSec(step.minUsersCount.get)
-          .to(step.maxUsersCount.get)
-          .during(getDuration(step.duration.get))
-      case _ =>
-        throw new IllegalArgumentException(
-          s"Invalid open model: ${step.action}"
-        )
-    }
-  }
-
-  def populationForStage(
-      scn: ScenarioBuilder,
-      steps: List[Step]
-  ): PopulationBuilder = {
-    logger.info(s"Building population for '${scn.name}' scenario")
-    if (steps.isEmpty) {
-      throw new IllegalArgumentException(
-        s"'${scn.name}': injection steps must be defined"
-      )
-    }
-    val models = steps
-      .map(step =>
-        modelsMap.getOrElse(
-          step.action,
-          throw new IllegalArgumentException(
-            s"Injection model '${step.action}' is not supported by GenericJourney"
-          )
-        )
-      )
-      .distinct
-    if (models.length > 1) {
-      throw new IllegalArgumentException(
-        s"'${scn.name}': closed and open models shouldn't be used in the same stage, fix scalabilitySetup:\n${modelsMap.toString()}"
-      )
-    }
-    models.head match {
-      case "open"   => scn.inject(steps.map(step => openModelStep(step)))
-      case "closed" => scn.inject(steps.map(step => closedModelStep(step)))
-      case _ =>
-        throw new IllegalArgumentException("Unknown step model")
-    }
-  }
 
   def testDataLoader(
       testData: TestData,
@@ -298,37 +89,45 @@ class GenericJourney extends Simulation {
       // Gatling automatically follow redirects in case of 301, 302, 303, 307 or 308 response status code
       // Disabling this behavior since we run the defined sequence of requests
       .disableFollowRedirect
-  private val steps = if (journey.needsAuthentication()) {
+
+  // Building sequence of http requests from journey file
+  private val httpSteps = if (journey.needsAuthentication()) {
+    // set 'Cookie' header with valid sid value to authenticate request
+    config.setCookieHeader = true
+
     /**
       * It does not make any difference to use unique cookie for individual user (tcp connection), unless we test Kibana
       * security service. Taking it into account, we create a single session and share it.
       */
     val cookiesLst = kbnClient.generateCookies(1)
     val circularFeeder = Iterator
-      .continually(cookiesLst.map(i => Map("sidValue" -> i)))
+      .continually(cookiesLst.map(i => Map("sid" -> i)))
       .flatten
     feed(circularFeeder)
-      .exec(session => session.set("Cookie", session("sidValue").as[String]))
-      .exec(scenarioSteps(journey, config))
-  } else exec(scenarioSteps(journey, config))
+      .exec(JourneyBuilder.buildHttpSteps(journey.streams, config))
+  } else {
+    exec(JourneyBuilder.buildHttpSteps(journey.streams, config))
+  }
 
-  private val warmupScenario = scenarioForStage(
-    steps,
+  private val warmupScenario = JourneyBuilder.buildScenario(
+    httpSteps,
     s"warmup for ${journey.journeyName} ${config.version}"
   )
-  private val testScenario = scenarioForStage(
-    steps,
+  private val testScenario = JourneyBuilder.buildScenario(
+    httpSteps,
     s"test for ${journey.journeyName} ${config.version}"
   )
 
   setUp(
-    populationForStage(warmupScenario, journey.scalabilitySetup.warmup)
+    JourneyBuilder
+      .buildPopulation(warmupScenario, journey.scalabilitySetup.warmup)
       .protocols(httpProtocol)
       .andThen(
-        populationForStage(testScenario, journey.scalabilitySetup.test)
+        JourneyBuilder
+          .buildPopulation(testScenario, journey.scalabilitySetup.test)
           .protocols(httpProtocol)
       )
-  ).maxDuration(getDuration(journey.scalabilitySetup.maxDuration))
+  ).maxDuration(JourneyBuilder.getDuration(journey.scalabilitySetup.maxDuration))
 
   before {
     if (journey.testData.isDefined) {
